requirements.txt
Flask==2.2.2
gql==3.4.0
python-dotenv==0.21.0
gunicorn==20.1.0